add manual approval

+++ Jenkinsfile
@@ -1,4 +1,6 @@
 node {
+    def SKIP_PROD = true;
+
     docker.image('python:2-alpine').inside {
         stage('Build') {
             checkout scm
@@ -12,10 +14,20 @@ node {
             junit 'test-reports/results.xml'
         }
     }
-    
-    stage('Deliver') {
-        checkout scm
-        sh 'docker run --rm -v $(pwd)/sources:/src cdrx/pyinstaller-linux:python2 \'pyinstaller -F add2vals.py\''
-        archiveArtifacts 'sources/dist/add2vals'
+
+    def approve
+    stage("Approval") {
+        approve = input(message: 'Lanjutkan ke tahap Deploy?', timeout: 1)
     }
+    
+    stage('Deploy') {
+        if (approve == 'proceed') {
+            checkout scm
+            sh 'docker run --rm -v $(pwd)/sources:/src cdrx/pyinstaller-linux:python2 \'pyinstaller -F add2vals.py\''
+            archiveArtifacts 'sources/dist/add2vals'
+        } else {
+            echo "Skip deploy"
+        }
+    }
+    
 }
